Change bills page link
